+++ src/main/java/com/revature/jukebox/App.java
@@ -1,10 +1,11 @@
 package com.revature.jukebox;
 
+import java.util.List;
+
 public class App {
     public static void main(String[] args){
-        //System.out.println("Hello");
         MusicRepository musicRepository = new MusicRepository("data.csv");
-        //MusicService musicService = new MusicService(musicRepository);
-        //MusicServer musicServer = new MusicServer(musicService);
+        MusicService musicService = new MusicService(musicRepository);
+        MusicServer musicServer = new MusicServer(musicService);
     }
 }