+++ src/main/java/com/revature/jukebox/IMusicRepoHandler.java
@@ -1,0 +1,6 @@
+package com.revature.jukebox;
+
+public interface IMusicRepoHandler{
+    default void monthReturn(){};
+    default void setID(){};
+}