+++ src/main/java/com/revature/jukebox/Month.java
@@ -1,0 +1,16 @@
+package com.revature.jukebox;
+
+public enum Month {
+    JANUARY,
+    FEBRUARY,
+    MARCH,
+    APRIL,
+    MAY,
+    JUNE,
+    JULY,
+    AUGUST,
+    SEPTEMBER,
+    OCTOBER,
+    NOVEMBER,
+    DECEMBER;
+}