+++ src/main/java/com/revature/jukebox/MusicRepoDeclarations.java
@@ -1,0 +1,10 @@
+package com.revature.jukebox;
+
+public record MusicRepoDeclarations() {
+    static Integer position;
+    static Integer streams;
+    static Integer year;
+    static Integer month;
+    static Integer day;
+    static String trackName, artist, genre;
+}