+++ CHANGELOG.md
@@ -7,4 +7,16 @@ Friday March 4th 2022
 ```aidl
 10:03 AM
 ```
--Adding in MusicRepository.java and App.java
+- Adding in MusicRepository.java and App.java
+
+```aidl
+10:43 AM
+```
+- Changed repository to load in all the files
+
+```aidl
+Monday March 7th 2022
+2:42 AM
+```
+- Working on the Interface and Records to handle MusicRepository
+- Set up the HTTP Servlet